% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allometric_equations.r
\name{extra_d4_funs}
\alias{extra_d4_funs}
\title{Allometric equations requiring Extra.D4 diameter}
\format{
A list of functions.
}
\usage{
extra_d4_funs
}
\description{
A list of allometric equation functions that require the
\code{Extra.D4} (basal diameter) measurement rather than the
standard \code{D4} diameter. When one of these functions is passed
as the \code{AGBFun} argument to \code{\link{CalcAGB}} or any of
the \code{SummaryAGWP} family of functions, the diameter column is
automatically switched from \code{"D4"} to \code{"Extra.D4"}.
}
\details{
To register a new allometric equation that requires \code{Extra.D4},
append it to this list.
}
\seealso{
\code{\link{AGBChv14}}, \code{\link{CalcAGB}}
}
