% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_simplify.R
\name{hd.simplify}
\alias{hd.simplify}
\title{Select height-diameter allometry parameters}
\usage{
hd.simplify(height.data, param.type = "Best")
}
\arguments{
\item{height.data}{Object returned by \code{\link{local.heights}}}

\item{param.type}{Level at which to return height-diameter parameters. One of "Best" (defualt), "ClusterF","BioRF" or "ContinentF}
}
\description{
Selects height-diameter allometry parameters at the desired level from the object returned by \code{\link{local.heights}}. Height-diameter allometric models are fitted at different levels of the dataset (e.g. cluster, biogeographic region) and can be either selected as the finest-scale available (\code{param.type}="Best", default) or at a specified level
}
