% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_mod.r
\name{height.mod}
\alias{height.mod}
\title{Estimate tree height from a height-diameter model}
\usage{
height.mod(
  dbh,
  data,
  model = "Model",
  a = "a_par",
  b = "b_par",
  c = "c_par",
  d = "d_par",
  height = "Height",
  ChaveE = "ChaveE",
  user.func = NULL
)
}
\arguments{
\item{dbh}{Numeric vector of diameters (DBH, mm).}

\item{data}{Dataframe containing columns reference in arguments below (e.g. model parameters).}

\item{model}{Character vector giving the type of height-diameter to use. Should be one of Weibull, Loglog, Env, Obs, MicMent, Hoerl, Gompertz, Polynomial or User. Obs uses measured heights.
User uses user defined function supplied to user.func.}

\item{a}{Numeric. Height-diameter model parameter.}

\item{b}{Numeric. Height-diameter model parameter.}

\item{c}{Numeric. Height-diameter model parameter.}

\item{d}{Numeric. Height-diameter model parameter.}

\item{height}{Numeric. Tree height (used when model is "Obs").}

\item{ChaveE}{Environmental stress parameter for height from Chave et al. 2014 Global Change Biology.}

\item{user.func}{User defined height-diameter function. Must use on or more of the parameters defined above.}
}
\value{
Numeric vector with estimated heights.
}
\description{
Function to estimate height from height-diameter model parameters
}
\author{
Martin Sullivan
}
