% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md}
\alias{md}
\title{Example plot metadata}
\format{
A data frame with 2 rows and 25 variables:
  \describe{
    \item{PlotCode}{Character plot code (e.g. \code{"JEN-11"}).}
    \item{PlotID}{Unique numeric plot identifier.}
    \item{PlotViewID}{Unique numeric plot-view identifier.}
    \item{Country}{Country name.}
    \item{Continent}{Continent name.}
    \item{Altitude}{Altitude above sea level (m).}
    \item{LatitudeDecimal}{Latitude in decimal degrees.}
    \item{LongitudeDecimal}{Longitude in decimal degrees.}
    \item{PlotArea}{Plot area (ha).}
    \item{ClusterName}{Name of the plot cluster.}
    \item{ClusterID}{Numeric cluster identifier.}
    \item{ForestMoistureName}{Forest moisture class name
      (e.g. \code{"Moist"}).}
    \item{ForestMoistureID}{Numeric forest moisture class
      identifier.}
    \item{ForestEdaphicName}{Forest edaphic type name
      (e.g. \code{"Terra Firma"}).}
    \item{ForestEdaphicID}{Numeric forest edaphic type
      identifier.}
    \item{ForestEdaphicHeightName}{Forest edaphic type name
      used for height-diameter allometry
      (e.g. \code{"Terra Firma"}).}
    \item{ForestEdaphicHeightID}{Numeric forest edaphic type
      identifier used for height-diameter allometry.}
    \item{ForestElevationName}{Forest elevation class name
      (e.g. \code{"Lowland"}).}
    \item{ForestElevationID}{Numeric forest elevation class
      identifier.}
    \item{ForestElevationHeightName}{Forest elevation class
      name used for height-diameter allometry
      (e.g. \code{"Lowland"}).}
    \item{ForestElevationHeightID}{Numeric forest elevation
      class identifier used for height-diameter allometry.}
    \item{BiogeographicalRegionName}{Biogeographical region
      name (e.g. \code{"Western Amazon"}).}
    \item{BiogeographicalRegionID}{Numeric biogeographical
      region identifier.}
    \item{RegionName}{Broad region name
      (e.g. \code{"Amazonia, W"}).}
    \item{AllometricRegionID}{Allometric region identifier,
      used to look up height-diameter parameters in
      \code{\link{WeibullHeightParameters}}.}
  }
}
\source{
ForestPlots.net Query Library > Basic plot information >
  Plot Information for R Package.
  \url{https://www.forestplots.net}
}
\usage{
md
}
\description{
Plot-level metadata for two ForestPlots.net plots (Ghana and Peru),
downloaded from the ForestPlots.net Query Library. This dataset is
provided as an example for use with \code{\link{mergefp}}.
}
\seealso{
\code{\link{trees}}, \code{\link{wd}}, \code{\link{mergefp}}
}
\keyword{datasets}
