% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomasaFP.R
\name{mergefp}
\alias{mergefp}
\alias{CleaningCensusInfo}
\title{Read and merge data downloaded from ForestPlots.net}
\usage{
mergefp(trees, md, wd)

CleaningCensusInfo(dfmerged)
}
\arguments{
\item{trees}{an individual tree data file downloaded from the "Advanced Search". Can be a dataframe or a character indicating the file path.}

\item{md}{a metadata file downloaded from the "Query Library". Can be a dataframe or a character indicating the file path.}

\item{wd}{a wood density file, with the wood density for each individual tree by PlotViewID downloaded from the "Query Library". Can be a dataframe or a character indicating the file path.}

\item{dfmerged}{Merged dataframe returned by \code{mergefp}.}
}
\description{
\code{mergefp} merges the 3 datasets needed for estimating biomass: Census Data (trees),
Metadata (md), wood density from each individual tree (wd). The three datasets have to be donwloaded from ForestPlots.net.
This function can use data frames and as well as file paths as input objects.
}
\references{
Lopez-Gonzalez G, Lewis SL, Burkitt M. and Phillips OL. 2011. ForestPlots.net: a web application and research tool to manage and analyse tropical forest plot data. Journal of Vegetation Science 22: 610–613. doi: 10.1111/j.1654-1103.2011.01312.x

Chave J, Coomes DA, Jansen S, Lewis SL, Swenson NG, Zanne AE. 2009. Towards a worldwide wood economics spectrum. Ecology Letters 12(4): 351-366. http://dx.doi.org/10.1111/j.1461-0248.2009.01285.x

Zanne AE, Lopez-Gonzalez G, Coomes DA et al. 2009. Data from: Towards a worldwide wood economics spectrum. Dryad Digital Repository. http://dx.doi.org/10.5061/dryad.234
}
\author{
Gabriela Lopez-Gonzalez, Tim Baker, Martin Sullivan
}
