% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryAGWP_Kohyama.R
\name{SummaryAGWP_Kohyama}
\alias{SummaryAGWP_Kohyama}
\title{Biomass gains and losses incorporating Kohyama et al. 2019 census interval correction}
\usage{
SummaryAGWP_Kohyama(
  xdataset,
  AGBEquation,
  dbh = "D4",
  height.data = NULL,
  palm.eq = TRUE,
  treefern = TRUE
)
}
\arguments{
\item{xdataset}{Object returned by \code{mergefp}.}

\item{AGBEquation}{Allometric equation function to use when estimating AGB, for example \code{\link{AGBChv14}}.}

\item{dbh}{Name of column containing diameter data. Default is "D4".}

\item{height.data}{Dataframe containing model type and parameters for height diameter model. These are matched on to the xdataset object at tree level. If NULL (default), then regional height-diameter equations (from Felpaush et al. 2012) are used.}

\item{palm.eq}{Logical. If TRUE the family level diameter based equation from Goodman et al 2013 is used to estimate AGB of monocots. Otherwise AGB of monocots is estimated using the allometric equation supplied to AGBEquation.}

\item{treefern}{Logical. If TRUE the height based equation from Tiepolo et al. 2002 is used to estimate the biomass of treeferns. If FALSE biomass is esimated using the allometric equation supplied to AGBEquation.}
}
\value{
Dataframe with AGWP and biomass losses for each census interval. NA for census 1. Both AGWP and losses incorporate unobserved growth, and are in units of Mg/ha/year.
}
\description{
Function to estimate AGWP and losses following equations 18 and 19 of Kohyama et al. 2019 Forest Ecology and Management. Acts as a wrapper for \code{turnover}.
}
\references{
Kohyama et al. 2019 Estimating net biomass production and loss from repeated measurements of trees in forests and woodlands: Formulae, biases and recomendations. Forest Ecology and Management 433: 729-740.
}
\author{
Martin Sullivan. Turnover function by T.S. Kohyama, T.I. Kohyama and D. Sheil.
}
