% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KohyamaProdLoss.R
\name{turnover}
\alias{turnover}
\title{Calculating biomass production and loss from repeated censuses}
\usage{
turnover(
  x1,
  x2,
  xs1,
  intvl,
  area,
  subpop = NULL,
  simple = FALSE,
  annual = FALSE
)
}
\arguments{
\item{x1}{Numeric vector. Biomass of stems in first census (Mg)}

\item{x2}{Numeric vector. Biomass of stems in second census (Mg)}

\item{xs1}{Numeric vector. Initial biomass of stems that survived from first census to second census (Mg)}

\item{intvl}{Numeric. Census interval length (years).}

\item{area}{Numeric. Plot area (ha)}

\item{subpop}{Vector giving subpopulation membership of each stem}

\item{simple}{Logical. Only used to obtain the different outputs from Koyhama et al.}

\item{annual}{Logical. Only used to obtain the different outputs from Koyhama et al.}
}
\description{
Code provided in SI of Koyhama et al. 2019 for estimating biomass gains and losses. Used by SummaryAGWP_Kohyama wrapper.
}
\references{
Kohyama et al. 2019 Estimating net biomass production and loss from repeated measurements of trees in forests and woodlands: Formulae, biases and recomendations. Forest Ecology and Management 433: 729-740.
}
\author{
T.S. Kohyama, T.I. Kohyama and D. Sheil
}
