% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wd}
\alias{wd}
\title{Example wood density data}
\format{
A data frame with 1309 rows and 5 variables:
  \describe{
    \item{PlotID}{Unique numeric plot identifier.}
    \item{PlotCode}{Character plot code (e.g. \code{"CAP-09"}).}
    \item{PlotViewID}{Unique numeric plot-view identifier.}
    \item{TreeID}{Unique numeric stem identifier.}
    \item{WD}{Wood density (g/cm\eqn{^3}).}
  }
}
\source{
ForestPlots.net Query Library > Wood Density > Wood Density
  of Individual Trees. \url{https://www.forestplots.net}
}
\usage{
wd
}
\description{
Wood density values for individual tree stems from two
ForestPlots.net plots (Ghana and Peru), downloaded from the
ForestPlots.net Query Library. This dataset is provided as an
example for use with \code{\link{mergefp}}.
}
\seealso{
\code{\link{trees}}, \code{\link{md}}, \code{\link{mergefp}}
}
\keyword{datasets}
