% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trees}
\alias{trees}
\title{Example tree census data}
\format{
A data frame with 8099 rows and 68 variables. Key variables
  include:
  \describe{
    \item{PlotID}{Unique numeric plot identifier.}
    \item{PlotViewID}{Unique numeric plot-view identifier.}
    \item{Plot.Code}{Character plot code (e.g. \code{"JEN-11"}).}
    \item{TreeID}{Unique numeric stem identifier.}
    \item{Census.No}{Census number (integer).}
    \item{Census.Date}{Decimal year of census.}
    \item{Family}{Accepted plant family name.}
    \item{Genus}{Accepted genus name.}
    \item{Species}{Accepted species name.}
    \item{AllometricRegionID}{Allometric region identifier, used to
      look up height-diameter parameters in
      \code{\link{WeibullHeightParameters}}.}
    \item{D1, D2, D3, D4}{Diameter measurements at successive
      quality-control steps (mm). \code{D4} is the recommended
      diameter for biomass calculations.}
    \item{Extra.D4}{Basal (extra) diameter measurement (mm), used
      by equations in \code{\link{extra_d4_funs}}.}
    \item{POM}{Point of measurement (mm above ground).}
    \item{F1}{Recruitment flag: \code{"n"} indicates a new recruit.}
    \item{F4}{Point-of-measurement change flag: \code{"6"} indicates
      a POM change.}
    \item{Height}{Tree height (m), where measured.}
  }
}
\source{
ForestPlots.net Advanced Search.
  \url{https://www.forestplots.net}
}
\usage{
trees
}
\description{
Individual tree stem measurements from two ForestPlots.net plots
(one in Ghana, one in Peru), covering multiple census years. This
dataset is provided as an example for use with
\code{\link{mergefp}} and the \code{SummaryAGWP} family of
functions.
}
\seealso{
\code{\link{md}}, \code{\link{wd}}, \code{\link{mergefp}}
}
\keyword{datasets}
